Add files via upload
Student Mannagement System is a GUI based application developed using python. 
Database- SQLite.

Libraries used in this application.
Tkinter library is used for GUI.
Data visualization of marks is done using Matplotlib.
Openweather API is used to display the temperature of the city in which user is located.
Beautifulsoup4 library is used for data extraction. Quote of the day is extracted from brainyquote website.
Requests library is used to get the ciy name and region name of user.